* Add icon.png in all module * Configuration module ** Configuration module form in the jimw.xml <configuration></configuration> node ** Edit and save configuration module ** Add icon in the module list to configure the module ** Add example configuration form in the blog module * Remove some debug display * Favicon : ** Add the default route favicon.ico ** Add automatically the meta favicon ** Use the favicon in the site path, then in the templace path, and finally in the public/common path ** Use favicon.png first then favicon.ico ** Default jimw favicon.ico * Common : ** Add helper common($file) to access to common file (automatically search file in common module directory and copy it in the common directory) ** Move pdf and print icon to the article's common directory ** Move mootools.js and gallery.js to the gallery's common directory * Correct the bug with the request URL witch can't allow to have GET var with a route * Add comments * Tree : ** Add the RecursiveIterator to the Jimw_Site_Tree_Rowset class ** Use the recursiveIterator in the default/tree/list and in the Tree Element form ** Tree Element form can be filtering to just display one module's tree * User : ** Add the param field "information" to allow to save other information in the user profile ** Correct error with date : must use $this->getFormatedDate(Zend_Date::now()) to set date * Correct error in the admin URL helper * Correct bug in debug mode with Link module because of the cache page cancel().

+++ trunk/jimw/lib/Jimw/Admin/Controller/ModuleController.php
@@ -28,7 +28,7 @@ class ModuleController extends Jimw_Admin_Action
         $modules = new Jimw_Site_Module();
         $modules_list = $modules->fetchAll();
         foreach ($modules_list as $module) {
-            $list[$module->path] = array('path' => $module->path , 'id' => $module->id , 'name' => $module->name , 'version' => $module->version , 'author' => $module->author , 'comment' => $module->comment , 'activ' => true);
+            $list[$module->path] = array('path' => $module->path , 'name' => $module->name , 'version' => $module->version , 'author' => $module->author , 'comment' => $module->comment , 'activ' => true);
         }
         $dir = new DirectoryIterator(JIMW_REP_MODULE);
         if ($dir->valid()) {
@@ -42,19 +42,58 @@ class ModuleController extends Jimw_Admin_Action
         ksort($list);
         $this->view->modules_list = $list;
     }
+
     public function disableAction ()
     {
         $modules = new Jimw_Site_Module();
         $trees = new Jimw_Site_Tree();
-        $module = $modules->find($this->_request->id)->current();
-        $list = $trees->fetchAll($trees->select()->where('module_path = ?', $module->path));
+        $module_path = $this->getRequest()->module_path;
+        $module = $modules->fetchRow($modules->select()->where('module_path = ?', $module_path));
+        if ($module == null) {
+            throw new Jimw_Admin_Exception('This module didn\'t exist');
+        }
+        $list = $trees->fetchAll($trees->select()->where('module_path = ?', $module_path));
         if (count($list) == 0)
-            $module->delete();
+        $module->delete();
         else {
             $this->_helper->getHelper('FlashMessenger')->addMessage($this->_('This module is used by page, you must delete all page to desactivate this module'));
         }
         $this->_forward('list');
     }
+
+    public function configAction ()
+    {
+        $modules = new Jimw_Site_Module();
+        $trees = new Jimw_Site_Tree();
+        $request = $this->getRequest();
+        $module_path = $request->module_path;
+        $module = $modules->fetchRow($modules->select()->where('module_path = ?', $module_path));
+        if ($module == null) {
+            throw new Jimw_Admin_Exception('This module didn\'t exist');
+        }
+        if (!isset($module->xml->configuration)) {
+            $this->_helper->getHelper('FlashMessenger')->addMessage($this->_('This module hasn\'t configuration'));
+            $this->_forward('list');
+            return;
+        }
+        $form = new Jimw_Form($module->xml->configuration);
+        if ($request->isPost() && $form->isValid($request->getPost())) {
+            $module->config = $form->getValues();
+            $module->save ();
+            $this->_helper->getHelper('FlashMessenger')->addMessage($this->_('Save configuration successful'));
+            $this->_forward('list');
+        }
+        else {
+            $form->addSubmit();
+            $config = $module->config;
+            if (config != null) {
+                $form->populate($config);
+            }
+            $this->view->form = $form;
+            $this->render('form');
+        }
+    }
+
     public function enableAction ()
     {
         $modules = new Jimw_Site_Module();
